requirements.txt
# local package
-e .

# PyPI packages
beautifulsoup4==4.12.3
click==8.1.7
darts==0.29.0
dateparser==1.2.0
matplotlib==3.4.3
numpy==1.22.4
pandas==1.5.3
pydantic==2.8.2
pypdf==4.3.0
python-dotenv==1.0.1
scipy==1.14.0
tabula_py==2.9.0
torch==2.3.0
u8darts==0.29.0
yfinance==0.2.38
